LPS-189216 Fix 'Resend' button remains on page even if user accepts invitation
