+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/types.d.ts
@@ -408,3 +408,19 @@ type UserLogged = {
 	isCustomerAccount: boolean;
 	isPublisherAccount: boolean;
 };
+
+type AdditionalInfoBody = {
+	acceptInviteStatus: boolean;
+	accountGroupERC: string;
+	accountName: string;
+	emailOfMember: string;
+	id: number;
+	inviteURL: string;
+	inviterName: string;
+	mothersName: string;
+	r_accountEntryToUserAdditionalInfo_accountEntryId: number;
+	r_userToUserAddInfo_userId: string;
+	roles: string;
+	sendType: {key: string; name: string};
+	userFirstName: string;
+};
