+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/utils/api.ts
@@ -965,11 +965,8 @@ export async function getMyUserAditionalInfos(userId: number) {
 		{headers}
 	);
 
-	const response = await userAdditionalInfos.json();
+	return await userAdditionalInfos.json();
 
-	if (!response.acceptInviteStatus) {
-		return response;
-	}
 }
 
 export async function updateUserPassword(password: string, id: number) {
